Initial working completion.
This is far from complete. There is a problem that needs to be resolved,
and it is explained below.

We observe the type of the symbol that is one character to the left of
the cursor (before the '.' or '->'). This works for cases where the
in-progress document is syntactically valid. So where:

line k  : Jsonrpc.Client client;
            ...
line n  : client. <-- keypress
line n+1: client.reply (id, null);

This code obeys the grammar required by the parser, but it fails the
semantic analysis test because the semantic analyzer interprets the
first symbol of the second line ("client") as a member of the object on
the first line ("client"). So "client.client" returns a semantic error
because "client" is not a member of "Jsonrpc.Client". However, because
we got to the semantic analysis part, the compiler already generated
the type symbol for "client" on the first line, so we can introspect
its type and return completion results.

However, take this example:

line n  : client. <-- keypress
line n+1: return;

Here, the document is not syntactically valid (since "return" is a
keyword), so the parsing stage of the compiler fails, and we never get
to the semantic analysis stage. So we never generate type symbols for
"client", and thus our completion list is empty.